// src/utils/enums.ts
export enum CardAttribute {
    DARK = "DARK",
    DIVINE = "DIVINE",
    EARTH = "EARTH",
    FIRE = "FIRE",
    LIGHT = "LIGHT",
    WATER = "WATER",
    WIND = "WIND",
}

export enum CardType {
    NORMAL_MONSTER = "NORMAL_MONSTER",
    EFFECT_MONSTER = "EFFECT_MONSTER",
    FLIP_EFFECT_MONSTER = "FLIP_EFFECT_MONSTER",
    FUSION_MONSTER = "FUSION_MONSTER",
    GEMINI_MONSTER = "GEMINI_MONSTER",
    LINK_MONSTER = "LINK_MONSTER",
    NORMAL_TUNER_MONSTER = "NORMAL_TUNER_MONSTER",
    PENDULUM_EFFECT_MONSTER = "PENDULUM_EFFECT_MONSTER",
    PENDULUM_EFFECT_FUSION_MONSTER = "PENDULUM_EFFECT_FUSION_MONSTER",
    PENDULUM_EFFECT_RITUAL_MONSTER = "PENDULUM_EFFECT_RITUAL_MONSTER",
    PENDULUM_FLIP_EFFECT_MONSTER = "PENDULUM_FLIP_EFFECT_MONSTER",
    PENDULUM_NORMAL_MONSTER = "PENDULUM_NORMAL_MONSTER",
    PENDULUM_TUNER_EFFECT_MONSTER = "PENDULUM_TUNER_EFFECT_MONSTER",
    RITUAL_EFFECT_MONSTER = "RITUAL_EFFECT_MONSTER",
    RITUAL_MONSTER = "RITUAL_MONSTER",
    SKILL_CARD = "SKILL_CARD",
    SPELL_CARD = "SPELL_CARD",
    SPIRIT_MONSTER = "SPIRIT_MONSTER",
    SYNCHRO_MONSTER = "SYNCHRO_MONSTER",
    SYNCHRO_PENDULUM_EFFECT_MONSTER = "SYNCHRO_PENDULUM_EFFECT_MONSTER",
    SYNCHRO_TUNER_MONSTER = "SYNCHRO_TUNER_MONSTER",
    TOKEN = "TOKEN",
    TOON_MONSTER = "TOON_MONSTER",
    TRAP_CARD = "TRAP_CARD",
    TUNER_MONSTER = "TUNER_MONSTER",
    UNION_EFFECT_MONSTER = "UNION_EFFECT_MONSTER",
    XYZ_MONSTER = "XYZ_MONSTER",
    XYZ_PENDULUM_EFFECT_MONSTER = "XYZ_PENDULUM_EFFECT_MONSTER",
}

export enum CardRace {
    NONE = "NONE", // for ""
    // Monsters
    AQUA = "AQUA",
    BEAST = "BEAST",
    BEAST_WARRIOR = "BEAST_WARRIOR",
    CREATOR_GOD = "CREATOR_GOD", // hello holactie
    CYBERSE = "CYBERSE",
    DINOSAUR = "DINOSAUR",
    DIVINE_BEAST = "DIVINE_BEAST",
    DRAGON = "DRAGON",
    FAIRY = "FAIRY",
    FIEND = "FIEND",
    FISH = "FISH",
    INSECT = "INSECT",
    MACHINE = "MACHINE",
    PLANT = "PLANT",
    PSYCHIC = "PSYCHIC",
    PYRO = "PYRO",
    REPTILE = "REPTILE",
    ROCK = "ROCK",
    SEA_SERPENT = "SEA_SERPENT",
    SPELLCASTER = "SPELLCASTER",
    THUNDER = "THUNDER",
    WARRIOR = "WARRIOR",
    WINGED_BEAST = "WINGED_BEAST",
    WYRM = "WYRM",
    ZOMBIE = "ZOMBIE",

    // Spells/Trap
    NORMAL = "NORMAL",
    QUICK_PLAY = "QUICK_PLAY",
    RITUAL = "RITUAL",
    CONTINIOUS = "CONTINIOUS",
    COUNTER = "COUNTER",
    EQUIP = "EQUIP",
    FIELD = "FIELD",

    // CHARACTERS
    CHAR_ALEXIS_RHODES = "CHAR_ALEXIS_RHODES",
    CHAR_ANDREW = "CHAR_ANDREW",
    CHAR_ARKANA = "CHAR_ARKANA",
    CHAR_ASTER_PHOENIX = "CHAR_ASTER_PHOENIX",
    CHAR_AXEL_BRODIE = "CHAR_AXEL_BRODIE",
    CHAR_BASTION_MISAW = "CHAR_BASTION_MISAW",
    CHAR_BONZ = "CHAR_BONZ",
    CHAR_CHAZZ_PRINCET = "CHAR_CHAZZ_PRINCET",
    CHAR_CHRISTINE = "CHAR_CHRISTINE",
    CHAR_CHUMLEY_HIFFI = "CHAR_CHUMLEY_HIFFI",
    CHAR_DAVID = "CHAR_DAVID",
    CHAR_EMMA = "CHAR_EMMA",
    CHAR_ESPA_ROBA = "CHAR_ESPA_ROBA",
    CHAR_ISHIZU_ISHTAR = "CHAR_ISHIZU_ISHTAR",
    CHAR_ISHIZU = "CHAR_ISHIZU",
    CHAR_JADEN_YUKI = "CHAR_JADEN_YUKI",
    CHAR_JESSE_ANDERSO = "CHAR_JESSE_ANDERSO",
    CHAR_JOEY_WHEELER = "CHAR_JOEY_WHEELER",
    CHAR_KAIBA = "CHAR_KAIBA",
    CHAR_KEITH = "CHAR_KEITH",
    CHAR_LUMIS_UMBRA = "CHAR_LUMIS_UMBRA",
    CHAR_MAI_VALENTINE = "CHAR_MAI_VALENTINE",
    CHAR_ODION = "CHAR_ODION",
    CHAR_MAKO = "CHAR_MAKO",
    CHAR_PARADOX_BROTHERS = "CHAR_PARADOX_BROTHERS",
    CHAR_PEGASUS = "CHAR_PEGASUS",
    CHAR_SYRUS_TRUESDA = "CHAR_SYRUS_TRUESDA",
    CHAR_TEA_GARDNER = "CHAR_TEA_GARDNER",
    CHAR_TYRANNO_HASSL = "CHAR_TYRANNO_HASSL",
    CHAR_WEEVIL = "CHAR_WEEVIL",
    CHAR_YAMI_BAKURA = "CHAR_YAMI_BAKURA",
    CHAR_YAMI_MARIK = "CHAR_YAMI_MARIK",
    CHAR_YAMI_YUGI = "CHAR_YAMI_YUGI",
    CHAR_YUBEL = "CHAR_YUBEL",
    CHAR_YUGI = "CHAR_YUGI",
    CHAR_ZANE_TRUESDALE = "CHAR_ZANE_TRUESDALE",
}

export enum CardFrame {
    EFFECT = "EFFECT",
    EFFECT_PENDULUM = "EFFECT_PENDULUM",
    FUSION = "FUSION",
    FUSION_PENDULUM = "FUSION_PENDULUM",
    LINK = "LINK",
    NORMAL = "NORMAL",
    NORMAL_PENDULUM = "NORMAL_PENDULUM",
    RITUAL = "RITUAL",
    RITUAL_PENDULUM = "RITUAL_PENDULUM",
    SKILL = "SKILL",
    SPELL = "SPELL",
    SYNCHRO = "SYNCHRO",
    SYNCHRO_PENDULUM = "SYNCHRO_PENDULUM",
    TOKEN = "TOKEN",
    TRAP = "TRAP",
    XYZ = "XYZ",
    XYZ_PENDULUM = "XYZ_PENDULUM",
}